[SYNCOPE-956] Protecting delete when Implementation is in use + refining Pull correlation rules

+++ core/logic/src/main/java/org/apache/syncope/core/logic/ImplementationLogic.java
@@ -29,8 +29,15 @@ import org.apache.syncope.common.lib.types.ClientExceptionType;
 import org.apache.syncope.common.lib.types.ImplementationType;
 import org.apache.syncope.common.lib.types.StandardEntitlement;
 import org.apache.syncope.core.persistence.api.dao.DuplicateException;
+import org.apache.syncope.core.persistence.api.dao.ExternalResourceDAO;
 import org.apache.syncope.core.persistence.api.dao.ImplementationDAO;
 import org.apache.syncope.core.persistence.api.dao.NotFoundException;
+import org.apache.syncope.core.persistence.api.dao.NotificationDAO;
+import org.apache.syncope.core.persistence.api.dao.PlainSchemaDAO;
+import org.apache.syncope.core.persistence.api.dao.PolicyDAO;
+import org.apache.syncope.core.persistence.api.dao.RealmDAO;
+import org.apache.syncope.core.persistence.api.dao.ReportDAO;
+import org.apache.syncope.core.persistence.api.dao.TaskDAO;
 import org.apache.syncope.core.persistence.api.entity.Implementation;
 import org.apache.syncope.core.provisioning.api.data.ImplementationDataBinder;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -45,6 +52,27 @@ public class ImplementationLogic extends AbstractTransactionalLogic<Implementati
 
     @Autowired
     private ImplementationDAO implementationDAO;
+
+    @Autowired
+    private ReportDAO reportDAO;
+
+    @Autowired
+    private PolicyDAO policyDAO;
+
+    @Autowired
+    private ExternalResourceDAO resourceDAO;
+
+    @Autowired
+    private TaskDAO taskDAO;
+
+    @Autowired
+    private RealmDAO realmDAO;
+
+    @Autowired
+    private PlainSchemaDAO plainSchemaDAO;
+
+    @Autowired
+    private NotificationDAO notificationDAO;
 
     @PreAuthorize("hasRole('" + StandardEntitlement.IMPLEMENTATION_LIST + "')")
     public List<ImplementationTO> list(final ImplementationType type) {
@@ -102,6 +130,69 @@ public class ImplementationLogic extends AbstractTransactionalLogic<Implementati
             LOG.error("Could not find implementation '" + key + "'");
 
             throw new NotFoundException(key);
+        }
+
+        boolean inUse = false;
+        switch (implementation.getType()) {
+            case REPORTLET:
+                inUse = !reportDAO.findByReportlet(implementation).isEmpty();
+                break;
+
+            case ACCOUNT_RULE:
+                inUse = !policyDAO.findByAccountRule(implementation).isEmpty();
+                break;
+
+            case PASSWORD_RULE:
+                inUse = !policyDAO.findByPasswordRule(implementation).isEmpty();
+                break;
+
+            case ITEM_TRANSFORMER:
+                inUse = !resourceDAO.findByTransformer(implementation).isEmpty();
+                break;
+
+            case TASKJOB_DELEGATE:
+                inUse = !taskDAO.findByDelegate(implementation).isEmpty();
+                break;
+
+            case RECON_FILTER_BUILDER:
+                inUse = !taskDAO.findByReconFilterBuilder(implementation).isEmpty();
+                break;
+
+            case LOGIC_ACTIONS:
+                inUse = !realmDAO.findByLogicActions(implementation).isEmpty();
+                break;
+
+            case PROPAGATION_ACTIONS:
+                inUse = !resourceDAO.findByPropagationActions(implementation).isEmpty();
+                break;
+
+            case PULL_ACTIONS:
+                inUse = !taskDAO.findByPullActions(implementation).isEmpty();
+                break;
+
+            case PUSH_ACTIONS:
+                inUse = !taskDAO.findByPushActions(implementation).isEmpty();
+                break;
+
+            case PULL_CORRELATION_RULE:
+                inUse = !policyDAO.findByCorrelationRule(implementation).isEmpty();
+                break;
+
+            case VALIDATOR:
+                inUse = !plainSchemaDAO.findByValidator(implementation).isEmpty();
+                break;
+
+            case RECIPIENTS_PROVIDER:
+                inUse = !notificationDAO.findByRecipientsProvider(implementation).isEmpty();
+                break;
+
+            default:
+        }
+
+        if (inUse) {
+            SyncopeClientException sce = SyncopeClientException.build(ClientExceptionType.InUse);
+            sce.getElements().add("This implementation is in use");
+            throw sce;
         }
 
         implementationDAO.delete(key);
